2019-06-07
Login:
Still issues with login, although both sign up and login seem to be working. Sign up stores credentials in the database and login creates a JWT token.
-must proceed by making the token expire (cookies?)
-make certain routes unaccessable until logged in

Tabs:
Inventaire tab made with a inverted color table. Sending data from the table to the database will not be too hard. Also created a dynamic table to have the option to add extra categories for extra consideration

Routing:
Routing needs to be fixed so it stays on refresh. There are conflicts with the react and laravel routing.

Future:

-Styling to make the app more presentable
-Be able to make multipage forms for questionnaires
-progress bar for questionnaires?