2019-06-25
Page prints out the table with all the data as required, change the state so that once the table is completed it is displayed and an option is given to edit the blank spaces. Or better, leave blank spaces in the table where information is missing.

Domain
Figure out what is causing the 403 error

Moving forward
Make all data tables and fix the tab layout to make it more intuitive